app/requirements.txt
python-dotenv
requests
flask
flask-cors
telebot
openai
google-auth
gspread
